pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "deepseek-ai>=0.0.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.75.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "xlsxwriter>=3.2.5",
]
